change-summary/scripts/pyproject.toml
[project]
name = "change-summary"
version = "0.1.0"
description = "Collect and pre-process git change data for LLM interpretation"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.3",
]

[project.scripts]
change-summary = "change_summary.cli:main"

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = ["ruff", "pytest>=8.0", "pytest-cov"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-ra", "--strict-markers", "--tb=short"]
markers = ["integration: tests requiring a git repo (deselect with '-m not integration')"]
